requirements.txt
openai==0.27.0
requests==2.28.2
requests-oauthlib==1.3.1
html2text==2020.1.16
python-dotenv==0.21.1